Extract rover naming to rover.go
