+++ pkg/rove/rover.go
@@ -1,8 +1,11 @@
 package rove
 
 import (
+	"bufio"
 	"fmt"
 	"log"
+	"math/rand"
+	"os"
 	"time"
 
 	"github.com/google/uuid"
@@ -60,7 +63,7 @@ func DefaultRover() Rover {
 		Capacity:         10,
 		Charge:           10,
 		MaximumCharge:    10,
-		Name:             uuid.New().String(),
+		Name:             GenerateRoverName(),
 	}
 }
 
@@ -74,4 +77,37 @@ func (r *Rover) AddLogEntryf(format string, args ...interface{}) {
 			Text: text,
 		},
 	)
+}
+
+var wordsFile = os.Getenv("WORDS_FILE")
+var roverWords []string
+
+// GenerateRoverName generates a new rover name
+func GenerateRoverName() string {
+
+	// Try and load the rover words file
+	if len(roverWords) == 0 {
+		// Try and load the words file
+		if file, err := os.Open(wordsFile); err != nil {
+			log.Printf("Couldn't read words file [%s], running without words: %s\n", wordsFile, err)
+		} else {
+			defer file.Close()
+			scanner := bufio.NewScanner(file)
+			for scanner.Scan() {
+				roverWords = append(roverWords, scanner.Text())
+			}
+			if scanner.Err() != nil {
+				log.Printf("Failure during word file scan: %s\n", scanner.Err())
+			}
+		}
+	}
+
+	// Assign a random name if we have words
+	if len(roverWords) > 0 {
+		// Loop until we find a unique name
+		return fmt.Sprintf("%s-%s", roverWords[rand.Intn(len(roverWords))], roverWords[rand.Intn(len(roverWords))])
+	}
+
+	// Default to a unique string
+	return uuid.New().String()
 }
